refactor: extract BinaryArrayDecoder and PropertyValueParser from ASGDecoder
Reduces ASGDecoder.java from 765 lines to 310 lines (target was <500).

Extracted:
- BinaryArrayDecoder: binary vertex/int/double array deserialization (240 lines)
- PropertyValueParser: XML helpers, text parsing, property value resolution (384 lines)

ASGDecoder retains thin delegates for decodeVertexArrayInBinary,
decodeIntArrayInBinary, decodeDoubleArrayInBinary to preserve the
existing call-site contract (ASG facade and reflection-based tests).

Public API on ASG.java is unchanged. All 72 existing tests pass.
New characterization test: ASGDecoderExtractionTest.

Closes #603
